Continuing game class

+++ Android/app/src/main/java/User.java
@@ -4,17 +4,17 @@
 public class User {
 
     public String name;// User name
-    public short id;      // Name identifier
+    public int id;      // Name identifier
 
     public static User NO_USER = new User("NO_USER", Short.MAX_VALUE);
 
-    public User(String username,short id){
+    public User(String username,int id){
         this.name = username;
         this.id = id;
     }
 
     public void setID(String idStr){
-        id = Short.parseShort(idStr);
+        id = (int) Short.parseShort(idStr);
     }
     public String getNameFull(){
         return name+Integer.toHexString(id);
